The GUI Release
Several changes in this update. Moved all code into a couple of
different packages for organizational purposes. Added a simple GUI to
help users navigate the program, and to allow them to choose a WAV file
to view. Windows generated by the program will also be centered on the
screen for easier viewing.